pyproject.toml
[project]
name = "prep-hackaton-toctoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "arize-phoenix>=7.3.2",
    "numpy>=2.2.1",
    "openai>=1.59.6",
    "pydantic>=2.10.4",
    "pydantic-ai>=0.0.18",
    "pydantic-ai-slim[openai]>=0.0.18",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=3.3.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "nest-asyncio>=1.6.0",
]
